Move FileItems up
